src/main/kotlin/com/radiotelescope/controller/model/appointment/CelestialBodyAppointmentCreateForm.kt
package com.radiotelescope.controller.model.appointment

import com.google.common.collect.HashMultimap
import com.google.common.collect.Multimap
import com.radiotelescope.contracts.appointment.ErrorTag
import com.radiotelescope.contracts.appointment.create.CelestialBodyAppointmentCreate
import java.util.*

data class CelestialBodyAppointmentCreateForm(
        override val userId: Long?,
        override val startTime: Date?,
        override val endTime: Date?,
        override val telescopeId: Long?,
        override val isPublic: Boolean?,
        val celestialBodyId: Long?
) : CreateForm<CelestialBodyAppointmentCreate.Request>() {
    override fun toRequest(): CelestialBodyAppointmentCreate.Request {
        return CelestialBodyAppointmentCreate.Request(
                userId = userId!!,
                startTime = startTime!!,
                endTime = endTime!!,
                telescopeId = telescopeId!!,
                isPublic = isPublic!!,
                celestialBodyId = celestialBodyId!!
        )
    }

    fun validateRequest(): Multimap<ErrorTag, String>? {
        val errors = HashMultimap.create<ErrorTag, String>()

        if (userId == null)
            errors.put(ErrorTag.USER_ID, "Invalid user id")
        if (startTime == null)
            errors.put(ErrorTag.START_TIME, "Required field")
        if (endTime == null)
            errors.put(ErrorTag.END_TIME, "Required field")
        if (telescopeId == null)
            errors.put(ErrorTag.TELESCOPE_ID, "Required field")
        if (isPublic == null)
            errors.put(ErrorTag.PUBLIC, "Required field")
        if (celestialBodyId == null)
            errors.put(ErrorTag.CELESTIAL_BODY, "Required field")

        return if (errors.isEmpty) null else errors
    }
}